pyproject.toml
[tool.poetry]
name = "access_log_reading"
version = "0.1.0"
description = ""
authors = ["matheus_campos <matheus@mobi2buy.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.3"
progressbar2 = "^3.53.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
